fixed the reloading issue
